Re-organized code, updated version and main entry point
