+++ lib/StandardRectangle/Label.js
@@ -3,7 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
     value: true
 });
-exports.LabelRow = exports.Label = exports.BlankLabel = undefined;
+exports.LabelRow = exports.Label = exports.BlankInsert = undefined;
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
@@ -45,7 +45,7 @@ var InsertWrapper = function InsertWrapper(_ref) {
     );
 };
 
-var BlankLabel = exports.BlankLabel = function BlankLabel() {
+var BlankInsert = exports.BlankInsert = function BlankInsert() {
     return _react2.default.createElement(
         _react2.default.Fragment,
         null,
