+++ lib/StandardRectangle/Sheet.js
@@ -19,11 +19,11 @@ var _Label = require('./Label');
 
 var _Label2 = _interopRequireDefault(_Label);
 
-var _PrintMediaQuery = require('./PrintMediaQuery');
+var _schema = require('./schema');
+
+var _PrintMediaQuery = require('../PrintMediaQuery');
 
 var _PrintMediaQuery2 = _interopRequireDefault(_PrintMediaQuery);
-
-var _schema = require('../schema');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
