+++ lib/StandardRectangle/index.js
@@ -1,0 +1,68 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.Layout = exports.SheetPrintView = exports.SheetPreview = exports.Sheet = exports.LabelRow = exports.BlankInsert = exports.Label = undefined;
+
+var _Label = require('./Label');
+
+Object.defineProperty(exports, 'Label', {
+  enumerable: true,
+  get: function get() {
+    return _interopRequireDefault(_Label).default;
+  }
+});
+Object.defineProperty(exports, 'BlankInsert', {
+  enumerable: true,
+  get: function get() {
+    return _Label.BlankInsert;
+  }
+});
+Object.defineProperty(exports, 'LabelRow', {
+  enumerable: true,
+  get: function get() {
+    return _Label.LabelRow;
+  }
+});
+
+var _Sheet = require('./Sheet');
+
+Object.defineProperty(exports, 'Sheet', {
+  enumerable: true,
+  get: function get() {
+    return _interopRequireDefault(_Sheet).default;
+  }
+});
+Object.defineProperty(exports, 'SheetPreview', {
+  enumerable: true,
+  get: function get() {
+    return _Sheet.SheetPreview;
+  }
+});
+Object.defineProperty(exports, 'SheetPrintView', {
+  enumerable: true,
+  get: function get() {
+    return _Sheet.SheetPrintView;
+  }
+});
+
+var _schema = require('./schema');
+
+Object.keys(_schema).forEach(function (key) {
+  if (key === "default" || key === "__esModule") return;
+  Object.defineProperty(exports, key, {
+    enumerable: true,
+    get: function get() {
+      return _schema[key];
+    }
+  });
+});
+
+var _Layout = require('../Layout');
+
+var _Layout2 = _interopRequireDefault(_Layout);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var Layout = exports.Layout = (0, _Layout2.default)(_Sheet.SheetPreview, _Sheet.SheetPrintView);
