+++ lib/StandardRectangle/schema.js
