--- lib/components/index.js
@@ -1,55 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _Label = require('./Label');
-
-Object.defineProperty(exports, 'Label', {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_Label).default;
-  }
-});
-Object.defineProperty(exports, 'BlankInsert', {
-  enumerable: true,
-  get: function get() {
-    return _Label.BlankInsert;
-  }
-});
-Object.defineProperty(exports, 'LabelRow', {
-  enumerable: true,
-  get: function get() {
-    return _Label.LabelRow;
-  }
-});
-
-var _Layout = require('./Layout');
-
-Object.defineProperty(exports, 'Layout', {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_Layout).default;
-  }
-});
-
-var _PrintMediaQuery = require('./PrintMediaQuery');
-
-Object.defineProperty(exports, 'PrintMediaQuery', {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_PrintMediaQuery).default;
-  }
-});
-
-var _Sheet = require('./Sheet');
-
-Object.defineProperty(exports, 'Sheet', {
-  enumerable: true,
-  get: function get() {
-    return _interopRequireDefault(_Sheet).default;
-  }
-});
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
