+++ lib/index.js
@@ -3,27 +3,12 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
+exports.StandardRectangle = undefined;
 
-var _components = require('./components');
+var _StandardRectangle2 = require('./StandardRectangle');
 
-Object.keys(_components).forEach(function (key) {
-  if (key === "default" || key === "__esModule") return;
-  Object.defineProperty(exports, key, {
-    enumerable: true,
-    get: function get() {
-      return _components[key];
-    }
-  });
-});
+var _StandardRectangle = _interopRequireWildcard(_StandardRectangle2);
 
-var _schema = require('./schema');
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
-Object.keys(_schema).forEach(function (key) {
-  if (key === "default" || key === "__esModule") return;
-  Object.defineProperty(exports, key, {
-    enumerable: true,
-    get: function get() {
-      return _schema[key];
-    }
-  });
-});
+exports.StandardRectangle = _StandardRectangle;
