+++ src/StandardRectangle/index.js
@@ -1,0 +1,8 @@
+import { SheetPreview, SheetPrintView } from './Sheet';
+import LayoutFactory from '../Layout';
+
+export { default as Label, BlankInsert, LabelRow } from './Label';
+export { default as Sheet, SheetPreview, SheetPrintView } from './Sheet';
+export * from './schema';
+
+export const Layout = LayoutFactory(SheetPreview, SheetPrintView);
