+++ src/StandardRectangle/schema.js
