--- src/components/index.js
@@ -1,4 +1,0 @@
-export { default as Label, BlankInsert, LabelRow } from './Label';
-export { default as Layout } from './Layout';
-export { default as PrintMediaQuery } from './PrintMediaQuery';
-export { default as Sheet } from './Sheet';
